Reboot NexusPlayers with adb and shut them down if install fails
BUG=skia:6045

+++ infra/bots/recipe_modules/flavor/gn_android_flavor.py
@@ -2,6 +2,8 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
+from recipe_engine import recipe_api
+
 import default_flavor
 import subprocess
 
@@ -42,6 +44,34 @@ class GNAndroidFlavorUtils(default_flavor.DefaultFlavorUtils):
     if 'infra_step' not in kwargs:
       kwargs['infra_step'] = True
     return self._run(title, 'adb', *cmd, **kwargs)
+
+  # Waits for an android device to be available
+  def _wait_for_device(self):
+    self.m.python.inline('wait for device', """
+      import subprocess
+      import sys
+      import time
+
+      times = 0
+
+      while times < 30:
+        print 'Waiting for the device to be connected and ready.'
+        try:
+          output = subprocess.check_output(['adb', 'shell',
+                                            'getprop', 'sys.boot_completed'])
+          if '1' in output:
+            print 'Connected'
+            sys.exit(0)
+        except subprocess.CalledProcessError:
+          # no device connected/authorized yet
+          pass
+        time.sleep(5)
+
+      print 'Timed out waiting for device'
+      sys.exit(1)
+      """,
+      infra_step=True)
+
 
   def compile(self, unused_target, **kwargs):
     compiler      = self.m.vars.builder_cfg.get('compiler')
@@ -89,8 +119,12 @@ class GNAndroidFlavorUtils(default_flavor.DefaultFlavorUtils):
     self._run('ninja', ninja, '-C', self.out_dir)
 
   def install(self):
+    if 'NexusPlayer' == self.m.vars.builder_cfg.get('model'):
+      self._adb('rebooting device', 'reboot')
+      self._wait_for_device()
     self._adb('mkdir ' + self.device_dirs.resource_dir,
               'shell', 'mkdir', '-p', self.device_dirs.resource_dir)
+
 
   def cleanup_steps(self):
     if self._ever_ran_adb:
@@ -112,6 +146,13 @@ class GNAndroidFlavorUtils(default_flavor.DefaultFlavorUtils):
       """,
       args=[self.m.vars.skia_out.join(self.m.vars.configuration)],
       infra_step=True)
+
+    for f in self.m.run.failed_steps:
+      if isinstance(f, recipe_api.InfraFailure):
+        self._adb('shut down device to quarantine bot', 'shell', 'reboot', '-p')
+        break
+
+    if self._ever_ran_adb:
       self._adb('kill adb server', 'kill-server')
 
   def step(self, name, cmd, env=None, **kwargs):
